add context to modal
