+++ packages/leafygreen-provider/src/index.ts
@@ -3,6 +3,10 @@ export {
   default as PortalContextProvider,
   usePopoverPortalContainer,
 } from './PortalContext';
+export {
+  usePopoverContext,
+  default as PopoverProvider,
+} from './PopoverContext';
 export { useBaseFontSize } from './TypographyContext';
 
 export { default } from './LeafyGreenContext';
