+++ packages/modal/src/index.ts
@@ -1,5 +1,6 @@
 import Footer from './Footer';
 import Modal, { ModalSize, CloseIconColor } from './Modal';
+import ModalView from './ModalView';
 
-export { Footer, ModalSize, CloseIconColor };
+export { Footer, ModalSize, CloseIconColor, ModalView };
 export default Modal;
